+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/BaseIncrementalCompilationTest.kt
@@ -2,20 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.gradle.incremental
 
+import com.autonomousapps.kit.GradleBuilder.build
+import com.autonomousapps.kit.GradleBuilder.buildAndFail
 import com.autonomousapps.kit.GradleProject
 import com.autonomousapps.kit.Source
 import com.autonomousapps.kit.Subproject
-import dev.zacsweers.metro.gradle.KotlinToolingVersion
 import dev.zacsweers.metro.gradle.copy
 import dev.zacsweers.metro.gradle.resolveSafe
 import java.io.File
 import org.intellij.lang.annotations.Language
 
 abstract class BaseIncrementalCompilationTest {
-
-  // TODO 2.3.20-Beta1 seems to have an IC regression
-  protected val brokenKotlin2320Versions =
-    setOf("2.3.20-dev-7791", "2.3.20-Beta1").mapTo(mutableSetOf(), ::KotlinToolingVersion)
 
   protected val GradleProject.asMetroProject: MetroGradleProject
     get() = MetroGradleProject(rootDir)
@@ -42,20 +39,84 @@ abstract class BaseIncrementalCompilationTest {
     get() = mainReports.forGraph("AppGraph", "test_AppGraph_Impl")
 
   class Reports(val dir: File) {
+    val expectActualReports
+      get() = dir.resolveSafe("expectActualReports.csv").readText()
+
+    val lookups
+      get() = dir.resolveSafe("lookups.csv").readText()
+
+    val traceLog
+      get() = dir.resolveSafe("traceLog.txt").readText()
+
+    val log
+      get() = dir.resolveSafe("log.txt").readText()
+
+    val timings
+      get() = dir.resolveSafe("timings.csv").readText()
+
+    fun irHintsForScope(scopeFqName: String): String {
+      return dir.resolveSafe("discovered-hints-ir-$scopeFqName.txt").readText()
+    }
+
+    fun firHintsForScope(scopeFqName: String): String {
+      return dir.resolveSafe("discovered-hints-fir-$scopeFqName.txt").readText()
+    }
+
     fun forGraph(simpleName: String, implFqName: String): GraphReports {
       return GraphReports(dir, simpleName, implFqName)
     }
   }
 
+  // TODO shared model?
   class GraphReports(val reportsDir: File, val simpleName: String, val implFqName: String) {
-    val keysPopulated: Set<String> by lazy {
-      reportsDir.resolveSafe("keys-populated-$simpleName.txt").readLines().toSet()
+    private fun readFileLines(path: String, extension: String = "txt"): List<String> {
+      return reportsDir.resolveSafe("$path.$extension").readLines()
     }
-    val providerPropertyKeys: Set<String> by lazy {
-      reportsDir.resolveSafe("keys-providerProperties-$implFqName.txt").readLines().toSet()
+
+    private fun readFile(pathWithExtension: String): String {
+      return reportsDir.resolveSafe(pathWithExtension).readText()
     }
-    val scopedProviderPropertyKeys: Set<String> by lazy {
-      reportsDir.resolveSafe("keys-scopedProviderProperties-$implFqName.txt").readLines().toSet()
+
+    val keysPopulated
+      get() = readFileLines("keys-populated-$simpleName")
+
+    val providerPropertyKeys
+      get() = readFileLines("keys-providerProperties-$implFqName")
+
+    val scopedProviderPropertyKeys
+      get() = readFileLines("keys-scopedProviderProperties-$implFqName")
+
+    val deferred
+      get() = readFileLines("keys-deferred-$simpleName")
+
+    val dumpKotlinLike
+      get() = readFile("graph-dumpKotlin-$implFqName.kt")
+
+    val dump
+      get() = readFileLines("graph-dump-$implFqName")
+
+    val bindingContainers
+      get() = readFileLines("graph-dump-$implFqName")
+
+    val keysValidated
+      get() = readFileLines("keys-validated-$simpleName")
+
+    val keysUnused
+      get() = readFileLines("keys-unused-$simpleName")
+
+    val metadata
+      get() = readFile("graph-metadata-$implFqName.kt")
+
+    val parentUsedKeysAll
+      get() = readFile("parent-keys-used-all-$simpleName")
+
+    fun parentKeysUsedBy(extension: String) =
+      readFileLines("parent-keys-used-$simpleName-by-$extension.txt")
+
+    fun graphMetadata() {
+      // /graph-metadata/graph-test-AppGraph.json"
+      // /graph-metadata/graph-test-AppGraph2.json"
+      TODO()
     }
   }
 
@@ -74,8 +135,9 @@ abstract class BaseIncrementalCompilationTest {
     rootDir: File,
     source: Source,
     @Language("kotlin") content: String,
+    includeDefaultImports: Boolean = true,
   ) {
-    val newSource = source.copy(content)
+    val newSource = source.copy(content, includeDefaultImports)
     val filePath = "src/main/kotlin/${newSource.path}/${newSource.name}.kt"
     val projectPath = rootDir.resolve(this.name.removePrefix(":").replace(":", "/"))
     projectPath.resolve(filePath).writeText(newSource.source)
@@ -97,4 +159,22 @@ abstract class BaseIncrementalCompilationTest {
     val filePath = "src/main/kotlin/$packageDir/$fileName"
     rootDir.resolve(filePath).writeText(content)
   }
+
+  protected fun GradleProject.compileKotlin(task: String = "compileKotlin") =
+    compileKotlin(rootDir, task)
+
+  protected fun GradleProject.compileKotlinAndFail(task: String = "compileKotlin") =
+    compileKotlinAndFail(rootDir, task)
+
+  protected fun compileKotlin(
+    projectDir: File,
+    task: String = "compileKotlin",
+    vararg args: String,
+  ) = build(projectDir, *listOf(task, "--quiet", *args).toTypedArray())
+
+  protected fun compileKotlinAndFail(
+    projectDir: File,
+    task: String = "compileKotlin",
+    vararg args: String,
+  ) = buildAndFail(projectDir, *listOf(task, "--quiet", *args).toTypedArray())
 }