+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changelog
 
 - [IR] Never eagerly init graph extension impls (scoped or not).
 - [IR] Don't cache creator-less scoped graph extension impls in their parent graphs. This was initially implemented this way due to a misunderstanding to how Dagger generated subcomponents! Getters for graph extensions now always return new instances.
+- [IC] Record lookups of contributed classes when looking up hints from IR. Previously Metro only recorded a lookup of the generated hint function, which appears to not be enough for Kotlin 2.3.20.
 
 ### Changes